+++ tests/test_cache.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+class TestEnfoldCache:
+    def test_init(self):
+        pass